Added Loads player is still in progress

+++ Assets/Scripts/Persistance/GameSave.cs
@@ -11,7 +11,7 @@ namespace Persistance
         public float GameOverTimer;
 
         public float[] BattleRoyaleTimers;
-        public float BattleRoyaleTimerIndex;
+        public int BattleRoyaleTimerIndex;
         public int RowCount;
         public int ColCount;
         public bool Paused;
